losetup: fix null-terminating src string instead of dest after strncpy
The function loopback_setup() forces the null-termination of the string
representing the loopback device name after the device name xstrncpy by
forcing its last byte to 0.

Unfortunately, this operation:
- should null-terminate dest instead;
- is useless because is the last operation performed on src before
  free();
- if the size of src is smaller than the offset of the written zero, it
  causes the modification of an unknown area in the heap.

Drop the null-termination line to fix the issue: xstrcpy already exits
with a failure if the size of src is bigger than the the requested
number of bytes to copy and automatically null-terminates dest.

Signed-off-by: Alessio Balsini <balsini@android.com>